Refactor: 패킷 구조 표준화 및 Protocol 정의 개선
기존 상속 구조의 패킷 정의를 멤버 포함(Composition) 방식으로 교체
패킷 헤더 내 ID 타입을 uint16_t로 통일하여 C2S/S2C 범용성 확보
불필요한 패킷 기본 클래스를 제거하고 #pragma pack 기반 데이터 정렬 최적화

+++ Network/Client/Client.cpp
@@ -43,7 +43,7 @@ int main() {
 	std::cout << "Connected to echo server." << std::endl;
 
 	C2S_CHAT chatPacket;
-	chatPacket.header.id = C2S_PACKET_ID::CHAT;
+	chatPacket.header.id = static_cast<uint16_t>(C2S_PACKET_ID::CHAT);
 	chatPacket.header.size =
 		sizeof(PACKET_HEADER) + sizeof("Hello, Echo Server!");
 	strcpy_s(chatPacket.message, sizeof(chatPacket.message),
